requirements.txt
fastapi==0.112.2
guardrails-ai
langchain==0.2.15
langchain_community==0.2.14
langchain_core==0.2.36
langchain_groq==0.1.9
langchain_openai==0.1.23
litellm==1.44.11
pandas==1.5.3
pandasai==2.2.14
pydantic
python-dotenv==1.0.1
uvicorn==0.30.6
plotly
kaleido
uvicorn
python-multipart
openpyxl
psycopg2_binary
pymysql
seaborn
jsonify
flask
tabulate
aiofiles
cachetools